// backend/src/apis/ZA/model/update-customer-profile-request.ts
/**
 * All API
 * TODO: Add Description
 *
 * The version of the OpenAPI document: 1.0
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

export class UpdateCustomerProfileRequest {
    'firstName': string;
    'lastName': string;
    'mobileNumber': string;
    'email': string;
    'address': string;
    'city': string;
    'customerId': string;
    'israelIdNumber': string;

    static discriminator: string | undefined = undefined;

    static attributeTypeMap: Array<{ name: string; baseName: string; type: string }> = [
        {
            name: 'firstName',
            baseName: 'firstName',
            type: 'string',
        },
        {
            name: 'lastName',
            baseName: 'lastName',
            type: 'string',
        },
        {
            name: 'mobileNumber',
            baseName: 'mobileNumber',
            type: 'string',
        },
        {
            name: 'email',
            baseName: 'email',
            type: 'string',
        },
        {
            name: 'address',
            baseName: 'address',
            type: 'string',
        },
        {
            name: 'city',
            baseName: 'city',
            type: 'string',
        },
        {
            name: 'customerId',
            baseName: 'customerId',
            type: 'string',
        },
        {
            name: 'israelIdNumber',
            baseName: 'israelIdNumber',
            type: 'string',
        },
    ];

    static getAttributeTypeMap() {
        return UpdateCustomerProfileRequest.attributeTypeMap;
    }
}